refactor: throttle util 함수에서 hook으로 변경
