--- web/src/utils/throttle.ts
@@ -1,14 +1,0 @@
-export const throttle = <Args extends unknown[]>(
-  fn: (...args: Args) => void,
-  delay: number,
-): ((...args: Args) => void) => {
-  let lastCall = 0;
-
-  return (...args: Args) => {
-    const now = Date.now();
-    if (now - lastCall >= delay) {
-      lastCall = now;
-      fn(...args);
-    }
-  };
-};
